LayoutTests/accessibility/url-test-expected.txt


test

This test ensures that we can retrieve URLs for appropriate elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.url.indexOf('url-test1.html') != -1 is true
PASS element.url.indexOf('url-test2.html') != -1 is true
PASS element.url.indexOf('url-test3.png') != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
